Rename packet_types to frame_types
